pyproject.toml
[project]
name = "portal-htr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "dotenv>=0.9.9",
    "kraken>=7.0.0b7",
    "pillow>=12.1.1",
    "requests>=2.33.1",
    "transformers>=5.4.0",
]
